Place shared ability examples in testing_support for extensions to access.
