+++ core/spec/models/ability_spec.rb
@@ -1,6 +1,7 @@
 require 'spec_helper'
 require 'cancan/matchers'
-require 'spree/core/testing_support/bar_ability'
+require 'spree/testing_support/ability_helpers'
+require 'spree/testing_support/bar_ability'
 
 # Fake ability for testing registration of additional abilities
 class FooAbility
@@ -41,81 +42,6 @@ describe Spree::Ability do
     it 'should apply the registered abilities permissions' do
       Spree::Ability.register_ability(FooAbility)
       Spree::Ability.new(user).can?(:update, mock_model(Spree::Order, :id => 1)).should be_true
-    end
-  end
-
-  shared_examples_for 'access granted' do
-    it 'should allow read' do
-      ability.should be_able_to(:read, resource, token) if token
-      ability.should be_able_to(:read, resource) unless token
-    end
-
-    it 'should allow create' do
-      ability.should be_able_to(:create, resource, token) if token
-      ability.should be_able_to(:create, resource) unless token
-    end
-
-    it 'should allow update' do
-      ability.should be_able_to(:update, resource, token) if token
-      ability.should be_able_to(:update, resource) unless token
-    end
-  end
-
-  shared_examples_for 'access denied' do
-    it 'should not allow read' do
-      ability.should_not be_able_to(:read, resource)
-    end
-
-    it 'should not allow create' do
-      ability.should_not be_able_to(:create, resource)
-    end
-
-    it 'should not allow update' do
-      ability.should_not be_able_to(:update, resource)
-    end
-  end
-
-  shared_examples_for 'index allowed' do
-    it 'should allow index' do
-      ability.should be_able_to(:index, resource)
-    end
-  end
-
-  shared_examples_for 'no index allowed' do
-    it 'should not allow index' do
-      ability.should_not be_able_to(:index, resource)
-    end
-  end
-
-  shared_examples_for 'create only' do
-    it 'should allow create' do
-      ability.should be_able_to(:create, resource)
-    end
-
-    it 'should not allow read' do
-      ability.should_not be_able_to(:read, resource)
-    end
-
-    it 'should not allow update' do
-      ability.should_not be_able_to(:update, resource)
-    end
-
-    it 'should not allow index' do
-      ability.should_not be_able_to(:index, resource)
-    end
-  end
-
-  shared_examples_for 'read only' do
-    it 'should not allow create' do
-      ability.should_not be_able_to(:create, resource)
-    end
-
-    it 'should not allow update' do
-      ability.should_not be_able_to(:update, resource)
-    end
-
-    it 'should allow index' do
-      ability.should be_able_to(:index, resource)
     end
   end
